+++ packages/decap-cms-core/src/components/UI/Modal.js
@@ -63,6 +63,9 @@ export class Modal extends React.Component {
   };
 
   componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(Modal.propTypes, this.props, 'prop', 'Modal');
+
     ReactModal.setAppElement('#nc-root');
   }