+++ packages/decap-cms-core/src/types/redux.ts
@@ -250,6 +250,7 @@ export interface CmsFieldStringOrText {
   // This is the default widget, so declaring its type is optional.
   widget?: 'string' | 'text';
   default?: string;
+  visualEditing?: boolean;
 }
 
 export interface CmsFieldMeta {
@@ -332,6 +333,7 @@ export interface CmsCollection {
   delete?: boolean;
   editor?: {
     preview?: boolean;
+    visualEditing?: boolean;
   };
   publish?: boolean;
   nested?: {
@@ -545,6 +547,8 @@ export type EntryObject = {
   slug: string;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   data: any;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  i18n?: any;
   collection: string;
   mediaFiles: List<MediaFileMap>;
   newRecord: boolean;