+++ packages/decap-cms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.4](https://github.com/decaporg/decap-cms/compare/decap-cms@3.6.3...decap-cms@3.6.4) (2025-06-06)
+
+**Note:** Version bump only for package decap-cms
+
+## [3.6.3](https://github.com/decaporg/decap-cms/compare/decap-cms@3.6.2...decap-cms@3.6.3) (2025-05-15)
+
+**Note:** Version bump only for package decap-cms
+
 ## [3.6.2](https://github.com/decaporg/decap-cms/compare/decap-cms@3.6.1...decap-cms@3.6.2) (2025-02-13)
 
 **Note:** Version bump only for package decap-cms